fix: JpaRepository의 제네릭 Integer -> Long 변경

+++ src/main/java/com/equip/equiprental/rental/repository/RentalItemRepository.java
@@ -3,5 +3,5 @@ package com.equip.equiprental.rental.repository;
 import com.equip.equiprental.rental.domain.RentalItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentalItemRepository extends JpaRepository<RentalItem, Integer> {
+public interface RentalItemRepository extends JpaRepository<RentalItem, Long> {
 }
